Alpha NavoznikAttack

+++ Assets/Scripts/Base_Components/Attack_Base.cs
@@ -7,7 +7,7 @@ public class Attack_Base : MonoBehaviour
     public BodyPart Part = BodyPart.Hands;
 
     public float AttackRange = 30f;
-    public int Damage;
+    public float Damage;
     public AudioClip sfx;
 
     public float AttackCooldown;
